+++ data_processing.h
@@ -4,11 +4,11 @@
 
 #define FILTER_LENGTH 117  // Example length, adjust as needed
 
-double calculateMean(double array[], int length);
-double calculateStdDev(double array[], int length, double mean);
-void calculateZScores(double array[], int length, double zScores[]);
-int find_peaks(double data[], int size, int peaks[]);
-void smooth_data(double data[], int size, double smoothed_data[]);
-void firFilter(int* input, double* output, int input_length, double* coefficients);
+float calculateMean(float array[], int length);
+float calculateStdDev(float array[], int length, float mean);
+void calculateZScores(float array[], int length, float zScores[]);
+int find_peaks(float data[], int size, int peaks[]);
+void smooth_data(float data[], int size, float smoothed_data[]);
+void firFilter(float* input, float* output, int input_length, float* coefficients);
 
 #endif
